Add a Kimurai-based scraper
It yields the spider and an opportunity for every opportunity scraped.
This allows us to use spider.save_to. Opportunity.to_h is currently
the only way to get any info. HT to
https://github.com/dxw/marketplace_opportunity_scraper from which a
bunch of the parsing code is pinched.

It does require a hack in that we override the spider's self.config,
because otherwise that will be empty.

At present, run with scrape.rb, and it only saves to result.json.