Add rake twitter:sign_in
Uses developer env vars from .env to initiate an OAuth 1.0a auth
conversation.

This makes a URL that the @dosalerts user can visit to authorize the app
and redirects to a non-existent local "Sinatra" app with a querystring
containing an oauth_identifier. That identifier can be pasted into the
running rake task, and then the rake task prints the access token and
secret such that the app can publish on behalf of @dosalerts.

Also add twitter:post_example to test.