Don't pass on_opportunity around in config
We were trying to get around that fact that Kimurai's crawl!
creates new instances of our spider class, and so
consequently we had no good place to store our callback if it wasn't
in the config.

However, config is stored at the class level anyway, so we may as
well just have a class-level attribute at the cost of us
never being able to thread two spiders at the same time with different
blocks.