Add ::Opportunity#type and #role
These are (sort of) available for scraping. If there's a role on the
page called "Specialist role" it's of type "Digital specialists".
If there's a "Description of your participants" it's of type
"User research participants".

In every other case it's of type "Digital outcomes".

The casing mirrors that on the main site and is validated as such.